+++ packages/react-docs/CHANGELOG.md
@@ -4,3 +4,4 @@
 
     * Move example out into a separate sub-package
     * Consolidate all `console.log` calls into `logUtils` in the `@0xproject/utils` package (#452)
+    * Handle `reflection` type rendering so that anonymous function type declarations render properly