+++ packages/react-docs/CHANGELOG.md
@@ -1,6 +1,6 @@
 # CHANGELOG
 
-## v0.0.2 - _TBD, 2018_
+## v0.0.2 - _March 18, 2018_
 
     * Move example out into a separate sub-package
     * Consolidate all `console.log` calls into `logUtils` in the `@0xproject/utils` package (#452)